Moved tokenizers tests to a dedicated file
